moved the routes that deal with saving posts into there own router
