+++ server/routes/settings.router.js
@@ -2,7 +2,6 @@ const express = require('express');
 const pool = require('../modules/pool');
 const sourceFetch = require('../modules/rss.js');
 const router = express.Router();
-let sourceFeed = require('../modules/rss.js');
 
 router.get('/sources', (req, res) => {
     console.log('getting sources for settings');
